+++ packages/ui-components/docs/changelog.md
@@ -5,6 +5,7 @@ title: Changelog
 ## NEXT
 
 -   Fixed an issue where if a word was too long, such as in an url, the `Chat` message would overflow instead of wrap.
+-   Fixed an issue where if a newline character was added to a comment, on subsequent reloads it would show '/n' in the message.
 
 ## 1.7.1