+++ packages/ui-components/docs/changelog.md
@@ -2,15 +2,19 @@
 title: Changelog
 ---
 
+## NEXT
+
+-   `Chat` component now displays its messages content as `Markdown`.
+-   `Chat` component now displays next line breaks.
+-   `Textarea` component when `onComplete` prop is passed adds next line with shift + enter. 
+-   Added a new `Markdown` component.
+
 ## 1.8.0
 
 -   **Breaking** In `Chat` messages, changed `timestamp` prop to be divided into two `created_at` and `updated_at`.
 -   Added "(edited)" indicator to the messages which have been edited in the `Chat` component.
 -   `Chat` will now send a message when `Enter` is pressed by the user.
 -   `Chat` now shows a disabled state for the `Send` button if the message to be sent is empty.
--   `Chat` component now displays its messages content as `Markdown`.
--   `Chat` component now displays next line breaks.
--   `Textarea` component when `onComplete` prop is passed adds next line with shift + enter. 
 
 ## 1.7.2